Remove domain from most URLs
With the expection of the QR Code URL, this makes URLs not be dependant
on QRMIRROR_PREFIX, since the links will just use the same domain as the
current page. The QR Code URL needs to be standalone, so QRMIRROR_PREFIX
is still needed to form the full public URL in QR codes.